Separated rest controllers from business logic and also created logic that deletes post interactions from the post interaction table whenever a post is deleted. Soon this implementation will be used to delete post interactions from the table as well as posts from its table when they associate with a user who deletes their account
