+++ src/main/java/com/videopostingsystem/videopostingsystem/posts/interaction/PostInteractionRepository.java
@@ -1,6 +1,5 @@
 package com.videopostingsystem.videopostingsystem.posts.interaction;
 
-import com.videopostingsystem.videopostingsystem.posts.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -9,4 +8,5 @@ import java.util.List;
 @Repository
 public interface PostInteractionRepository extends JpaRepository<PostInteractions, String> {
     List<PostInteractions> findByUsers(String users);
+    List<PostInteractions> findByPostID(Long postID);
 }
