+++ src/main/java/com/videopostingsystem/videopostingsystem/posts/interaction/PostInteractionRequest.java
@@ -1,0 +1,24 @@
+package com.videopostingsystem.videopostingsystem.posts.interaction;
+
+import jakarta.servlet.http.HttpSession;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class PostInteractionRequest {
+
+    private final PostInteractionService postInteractionService;
+
+    public PostInteractionRequest(PostInteractionService postInteractionService){
+        this.postInteractionService = postInteractionService;
+    }
+
+
+    @PostMapping("/posts/{postId}")
+    public ResponseEntity<?> postInteraction(@PathVariable("postId") Long postId, @RequestBody PostInteractionModel postInteraction, HttpSession session){
+       return postInteractionService.postInteraction(postId, postInteraction, session);
+    }
+}
